Change type mapping structure
Type mapping is now structured within ozo/pg/types
