+++ include/ozo/ext/std.h
@@ -13,6 +13,7 @@
 #include <ozo/ext/std/nullptr_t.h>
 #include <ozo/ext/std/optional.h>
 #include <ozo/ext/std/shared_ptr.h>
+#include <ozo/ext/std/string.h>
 #include <ozo/ext/std/tuple.h>
 #include <ozo/ext/std/unique_ptr.h>
 #include <ozo/ext/std/vector.h>
