+++ include/ozo/pg/types.h
@@ -1,0 +1,21 @@
+#pragma once
+
+/**
+ * @defgroup group-type_system-pg-types PostgreSQL
+ * @ingroup group-type_system
+ * @brief PostgreSQL types reflections definitions.
+ *
+ * Contains definition of types are reflecting PostgreSQL built-in types.
+ */
+
+#include <ozo/pg/types/bool.h>
+#include <ozo/pg/types/bytea.h>
+#include <ozo/pg/types/char.h>
+#include <ozo/pg/types/float.h>
+#include <ozo/pg/types/integer.h>
+#include <ozo/pg/types/json.h>
+#include <ozo/pg/types/jsonb.h>
+#include <ozo/pg/types/name.h>
+#include <ozo/pg/types/oid.h>
+#include <ozo/pg/types/text.h>
+#include <ozo/pg/types/uuid.h>
