+++ include/ozo/pg/types/text.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <ozo/type_traits.h>
+#include <ozo/ext/std/string.h>
+
+namespace ozo::pg {
+using text = std::string;
+} // namespace ozo:pg
