+++ include/ozo/pg/types/uuid.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <ozo/ext/boost/uuid.h>
+
+namespace ozo::pg {
+using uuid = boost::uuids::uuid;
+} // namespace ozo::pg
