+++ tests/binary_deserialization.cpp
@@ -3,7 +3,7 @@
 #include <ozo/io/array.h>
 #include <ozo/io/recv.h>
 #include <ozo/ext/std.h>
-#include <ozo/ext/boost/uuid.h>
+#include <ozo/pg/types.h>
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
