+++ tests/integration/result_integration.cpp
@@ -2,6 +2,7 @@
 #include <ozo/io/recv.h>
 #include <ozo/ext/boost.h>
 #include <ozo/ext/std.h>
+#include <ozo/pg/types.h>
 #include <ozo/io/array.h>
 #include <ozo/io/composite.h>
 #include <ozo/shortcuts.h>
